Chat with OpenAI's GPT-3.5-turbo model HEJ MED DIG

This function takes a data frame of study details, constructs a prompt for each study, and sends it to OpenAI's GPT-3.5-turbo model for a response. The function returns a tibble with the original study details and the model's response.

Usage

chat(data, prompt, model = "gpt-3.5-turbo")

Arguments

data

A data frame containing the study details. Each row should represent a study and must have 'title' and 'abstract' columns.

prompt

A string that will be prepended to the study details to form the prompt for the model.

model

A string specifying the model to use. Default is "gpt-3.5-turbo".

Value

A tibble with the original study details and the model's response.

Examples

if (FALSE) { # \dontrun{
prompt <- "Is this study about a Functional Family Therapy (FFT) intervention?"
loaded_objects <- load("./data/filges2015_dat_modified.rda")
data <- get(loaded_objects)
chat(data, prompt)
} # }